June 2022 Unity project update
Follow Unity daily updates at Miro:
https://miro.com/app/board/o9J_l6Ti6DY=/

Additional file added
- pngs: Gigachad, Mr.White, Naked Snake, Gus Fringe
- Game Design Paper: 1) Inventory system and UI layout 2) Enemy spawning type

Unity Project System Update:
Before getting to what's new in this commit, I'd like to explain that most updates going on Elimination mode or Zeus Mode for short, it's the core gameplay mode of Math Conquest that need to be focused. More of that, the script of this kind of gameplay is out of my coding field, which leads to longer working on a single system with dynamic input.

- Basic control system of Exploration mode (3D puzzle) and Dungeon mode (2D side-scroll)
- Dialog system between characters (dynamic dialogs)
- Basic Calculation Question Generator that can be used in every mode
- Enemy Elimination mode  cycle action system (see how it works in Miro)
- Enemy Elimination mode Spawning system: has 4 setting types of enemy spawn type (completed by 2 type: move-in and waiting)
- Player Elimination mode Input system
- Enemy Status: control the enemy's HP

Currently in development and further on:
- Stage Event Control System in Elimination mode
- Dialog system that includes the background picture
- 2 types of enemy Elimination mode spawning types: fill-in and  pop-up