remove older Core project
The older core project had been removed due to the selection of project type during setup. The 2D project setup was selected and that led to some missing 3D features and false display of 3D object lighting. The solve the unintended problem, the new project with 3D setup was created, then packed all the files as Unity Package in the old one. Move them to a new setup named 3D Core.